fix: refactor hero and form section
