--- src/components/pages/landing/subscribe-form/index.js
@@ -1,3 +1,0 @@
-import SubscribeForm from './subscribe-form';
-
-export default SubscribeForm;
